+++ packages/braintree-integration/src/braintree-accelerated-checkout/braintree-accelerated-checkout-payment-initialize-options.ts
@@ -1,3 +1,5 @@
+import { BraintreeConnectStylesOption } from '@bigcommerce/checkout-sdk/braintree-utils';
+
 /**
  * A set of options that are required to initialize the Braintree Accelerated Checkout payment
  * method for presenting on the page.
@@ -13,16 +15,53 @@
  * service.initializePayment({
  *     methodId: 'braintreeacceleratedcheckout',
  *     braintreeacceleratedcheckout: {
- *         onInit: (renderPayPalComponentMethod) => renderPayPalComponentMethod('#container-id');
+ *         onInit: (renderPayPalComponentMethod) => renderPayPalComponentMethod('#container-id'),
+ *         styles: {
+ *              root: {
+ *                  backgroundColorPrimary: 'transparent',
+ *                  errorColor: '#C40B0B',
+ *                  fontFamily: 'Montserrat, Helvetica, Arial, sans-serif',
+ *              },
+ *              input: {
+ *                  borderRadius: '0.25rem',
+ *                  borderColor: '#9E9E9E',
+ *                  focusBorderColor: '#4496F6',
+ *              },
+ *              toggle: {
+ *                  colorPrimary: '#0F005E',
+ *                  colorSecondary: '#ffffff',
+ *              },
+ *              text: {
+ *                  body: {
+ *                      color: '#222222',
+ *                      fontSize: '1rem',
+ *                  },
+ *                  caption: {
+ *                      color: '#515151',
+ *                      fontSize: '0.875rem',
+ *                  },
+ *              },
+ *              branding: 'light',
+ *         },
  *     },
  * });
  * ```
  */
 export default interface BraintreeAcceleratedCheckoutPaymentInitializeOptions {
     /**
-     * The CSS selector of a container where the payment widget should be inserted into.
+     * Is a callback that takes the CSS selector of a container
+     * where the PayPal Connect form should be inserted into.
      */
     onInit?: (renderPayPalComponentMethod: (container: string) => void) => void;
+
+    /**
+     * Is a stylisation options for customizing PayPal Connect components
+     *
+     * Note: the styles for all Braintree Accelerated Checkout strategies should be the same,
+     * because they will be provided to PayPal library only for the first strategy initialization
+     * no matter what strategy was initialised first
+     */
+    styles?: BraintreeConnectStylesOption;
 }
 
 export interface WithBraintreeAcceleratedCheckoutPaymentInitializeOptions {