+++ packages/braintree-utils/src/index.ts
@@ -1,41 +1,8 @@
-export {
-    BraintreeModuleCreator,
-    BraintreeInitializationData,
-    BraintreeTokenizePayload,
-    BraintreeError,
-    BraintreeDataCollector,
-    BraintreePaypalCheckoutCreator,
-    BraintreePaypalCheckout,
-    BraintreePaypalSdkCreatorConfig,
-    BankAccountSuccessPayload,
-    BraintreeBankAccount,
-    BraintreeConnect,
-    BraintreeConnectAuthenticationState,
-    BraintreeConnectAddress,
-    BraintreeConnectVaultedInstrument,
-    BraintreeConnectCardComponentOptions,
-    BraintreeConnectCardComponent,
-    BraintreeConnectPhone,
-    BraintreeConnectProfileData,
-    onPaymentStartData,
-    StartPaymentError,
-    LocalPaymentsPayload,
-    LocalPaymentInstance,
-    BraintreeHostWindow,
-    GooglePayBraintreeSDK,
-    BraintreeClient,
-    BraintreeThreeDSecure,
-    BraintreeGooglePayThreeDSecure,
-} from './braintree';
-
+export * from './braintree';
 export { PaypalSDK, PaypalButtonOptions, PaypalAuthorizeData } from './paypal';
-
 export { default as BraintreeIntegrationService } from './braintree-integration-service';
-
 export { default as BraintreeScriptLoader } from './braintree-script-loader';
-
 export { default as isBraintreeError } from './is-braintree-error';
-
 export {
     getBraintreeConnectProfileDataMock,
     getConnectMock,
@@ -50,5 +17,5 @@ export {
     getClientMock,
     getThreeDSecureMock,
 } from './mocks/braintree.mock';
-
+export { getValidBraintreeConnectStyles } from './utils';
 export { BRAINTREE_SDK_STABLE_VERSION, BRAINTREE_SDK_ALPHA_VERSION } from './sdk-verison';