+++ packages/core/src/payment/strategies/clearpay/clearpay-script-loader.spec.ts
@@ -8,19 +8,14 @@ import ClearpayScriptLoader from './clearpay-script-loader';
 describe('ClearpayScriptLoader', () => {
     const scriptLoader = createScriptLoader();
     const clearpayScriptLoader = new ClearpayScriptLoader(scriptLoader);
-
-    beforeEach(() => {
-        jest.spyOn(scriptLoader, 'loadScript').mockReturnValue(Promise.resolve(new Event('load')));
-    });
+    const loadScript = jest.spyOn(scriptLoader, 'loadScript');
 
     it('loads widget script', () => {
         const method = getClearpay();
 
         clearpayScriptLoader.load(method);
 
-        expect(scriptLoader.loadScript).toHaveBeenCalledWith(
-            '//portal.clearpay.co.uk/afterpay-async.js',
-        );
+        expect(loadScript).toHaveBeenCalledWith('//portal.clearpay.co.uk/afterpay-async.js');
     });
 
     it('loads sandbox widget script if in test mode', () => {
@@ -28,7 +23,7 @@ describe('ClearpayScriptLoader', () => {
 
         clearpayScriptLoader.load(method);
 
-        expect(scriptLoader.loadScript).toHaveBeenCalledWith(
+        expect(loadScript).toHaveBeenCalledWith(
             '//portal.sandbox.clearpay.co.uk/afterpay-async.js',
         );
     });