+++ packages/core/src/payment/strategies/googlepay/googlepay.ts
@@ -82,14 +82,17 @@ export interface GooglePayHostWindow extends Window {
     google?: GooglePaySDK;
 }
 
+export interface CardInformation {
+    cardType: string;
+    lastFour: string;
+    lastTwo?: string;
+    bin?: string;
+}
+
 export interface TokenizePayload {
     nonce: string;
     tokenFormat?: string;
-    details: {
-        cardType: string;
-        lastFour: string;
-        lastTwo?: string;
-    };
+    details: CardInformation;
     description?: string;
     type: TokenizeType;
     binData?: {
@@ -152,6 +155,7 @@ export interface PaymentMethodData {
         cardInformation: {
             type: string;
             number: string;
+            bin: string;
         };
     };
 }