+++ packages/core/src/payment/strategies/klarnav2/klarnav2-script-loader.spec.ts
@@ -5,16 +5,11 @@ import KlarnaV2ScriptLoader from './klarnav2-script-loader';
 describe('KlarnaV2ScriptLoader', () => {
     const scriptLoader = createScriptLoader();
     const klarnav2ScriptLoader = new KlarnaV2ScriptLoader(scriptLoader);
-
-    beforeEach(() => {
-        jest.spyOn(scriptLoader, 'loadScript').mockReturnValue(Promise.resolve(true));
-    });
+    const loadScript = jest.spyOn(scriptLoader, 'loadScript');
 
     it('loads widget script', () => {
         klarnav2ScriptLoader.load();
 
-        expect(scriptLoader.loadScript).toHaveBeenCalledWith(
-            'https://x.klarnacdn.net/kp/lib/v1/api.js',
-        );
+        expect(loadScript).toHaveBeenCalledWith('https://x.klarnacdn.net/kp/lib/v1/api.js');
     });
 });